requirements.txt
asgiref==3.7.2
crispy-bootstrap5==0.7
defusedxml==0.7.1
Django==4.2.6
django-crispy-forms==2.1
django-filter==23.3
django-htmx==1.17.0
django-tables2==2.6.0
fonttools==4.45.1
fpdf2==2.7.6
Pillow==10.1.0
pypdf==3.17.1
sqlparse==0.4.4
